+++ forui_hooks/test/src/multi_value_notifier_hook_test.dart
@@ -16,8 +16,7 @@ void main() {
           builder: (context) {
             controller = useFMultiValueNotifier<int>();
             return FSelectGroup<int>(
-              controller: controller,
-              children: [
+              control: .managed(controller: controller), children: [
                 .checkbox(value: 0, label: const Text('0')),
                 .checkbox(value: 1, label: const Text('1')),
               ],
@@ -41,8 +40,7 @@ void main() {
           builder: (context) {
             controller = useFRadioMultiValueNotifier<int>();
             return FSelectGroup<int>(
-              controller: controller,
-              children: [
+              control: .managed(controller: controller), children: [
                 .radio(value: 0, label: const Text('0')),
                 .radio(value: 1, label: const Text('1')),
               ],